Medical Doctor Panels
I created panels classes for each type of section and then I built 2
test frames.  TestMedicalFrame is the first frame I built but I'm not
sure about the design so I designed a second MedicalMainFrame.   Once we
decide on which design we like better I can implement the
DoctorViewFrame and the NurseViewFrame for these panels by extending
whichever Frame we choose to go with and disabling and enabling
textfield etc.