New Mavenized Project
I had to create a subproject under guardians-md because Maven doesn't
like the - when structuring the packaged jar file.  I added all the
package and imports to all the files.  I created a DbConnDummy that
implements DBConn and fixed all the compilation errors for a clean
install of the package.  When you run mvn clean package  a /target dir
will be created and a guardians_md.jar file.  If you click on the jar
file it executes the application, which is set to display the Nursing
Panel for now.  This configuration will change in the pom file.