DatabaseConnection is Functional
I've completed work on making the DatabaseConnection functional. Let me
know if anything isn't working as expected or as you need it to. I
think my testing was fairly solid, but note that:
1. Currently Prescriptions and LabOrders cannot be edited once created.
2. More testing may be required for multiple prescriptions and lab
orders on one visit - I believe this is working but the code for this
isn't as clean as I'd like it to be, so something may have snuck by me.