Database test class
This class runs tests on the DatabaseConnection code to verify that it
functions as expected.

Once this file is included in your local repository, you will
experience a "BUILD FAILURE" message if you attempt to run a maven
build that includes tests without setting up your local MYSQL server.

If you have the MYSQL server set up but it is not connecting correctly,
it's most likely a port error or a username/password error. Check the
top of the DatabaseConnection file and look for differences with what
your mysql server settings are.