S3 Image Storage Functionality
Added functionality to store, retrieve, and delete images on an S3
cloud bucket. This can be accessed via an instance of the
S3ImageStorage class.

This commit also includes a test class (S3Test) for the image storage.

I also had to add a couple of dependencies to the maven project to make
this work, primarily the AWS SDK dependency. You may notice Maven
downloading more artifacts now when it builds/packages.