added new frame and implementation ResultsFrameView
The frame has the ability to show images or texts which allows for the
use of the same frame for all the lab order results that are either
Strings or images.

Tested the frame in an isolated fashion using the additional constructor
that allows for the boolean hasResultImage.  Was able to see the image
in the frame after setting the lab order using the MenuConfigurator.