Added property loading to MedicalConfigurator
The MedicalConfigurator class will now load properties from a
db.properties file in one of the following locations:
1. In the working directory of the .jar (used when executing packaged
the .jar file).
2. Packaged in test->resources (used only for running unit tests)
3. Packaged in main->resources on build
It will prioritize locations higher in this list.

An example db.properties file has been included in the top level github
repository. This example is for running the program locally.

I've also included a packaged jar which was packaged with a
db.properties to run on the AWS server. Because of Amazon security, I
cannot include the plaintext db.properties file which contains the AWS
secret access key.